Updated portrait picture and structure of contents
Swapped the Positions of the portrait picture and text and also replaced the picture with a professional one.

Removed a sentence in "about me" and also changed the wording of the second sentence to sound better.

restructured the contents so that "Projekte" (projects) comes before "Werdegang" (my cv).

Changed the order of the nav elements to match the structure of the site.

Added class "hidden" to li elments of "Werdegang" to apply scrol animations.

Restructured the description of my previous workplaces.